perf: speed up `EncodeSourceVersion` by multiple optimizations
1. Use `WalkDir` to only call `os.Stat` when needed (i.e. when access to file modification
time is needed). Previously, it was called on all entries.
This cut `EncodeSourceVersion` time in one of our private workspaces from 43s to 34s.

2. Skip walking cache directories completely by returning `filepath.SkipDir` when hitting
one. Previously, we just returned nil which meant we still had to walk everything in the
file tree.
This cut `EncodeSourceVersion` time in one of our private workspaces from 34s to ~nil.